refactor: moving message handling out of networkplayer
Server or client should handle message instead of player. Makes registering message simpler, no need to wait for player to Connect to register message

BREAKING CHANGE: RegisterHandler functions now exist on MessageHandler On Server and Client
