+++ Assets/Mirage/Runtime/MessageHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a9b600aef1886a64c82a2263b230f007
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
